Don't require proxy to be enabled for custom branding to be enabled
I realised firstly that:
- We need to be able to customise branding config in scenarios where packit-deploy
does not provide the proxy, and its provided otherwise, as is the set-up in Montagu.
Montagu handles the proxy in that case.
- But running packit-deploy in a proxy-less mode doesn't in fact make knowing the
app file locations complicated, since the Packit app Dockerfile specifies the location
for the built app to be copied to, so we can rely on that to be the case for both
proxy-less and non-proxy-less modes of packit-deploy.