Break up PackitConfig into smaller dataclasses.
This groups related functionality together and more or less mirrors the
structure of the configuration file. This is done to support multiple
instances of Packit, where bits like the Packit API, the database, ...
will have multiple instances.

At some point in the future it may be possible to replace a lot of the
boilerplate code with something like pydantic.