Simplify the way container names are handled.
We used to have a dictionary for all the container names.
This will become a bit unwieldy once we have many replicas of a
container. Instead the names are put as fields of the correspoding
container configuration objects, usually as constants.

Additionally stop putting the constellation container prefix in the
URLs. It turns out that constellation sets up networking such that these
are not needed.