Build the proxy image on the fly.
Previously the integration tests were always using the proxy image from
GHCR. This makes changes to the proxy difficult.

The new `build` option allows a path to a directory containing a Docker
file to be specified in place of a name and tag. We use it for the proxy
everywhere.